refactor: consolidate lexicon tests
