+++ tests/test_lexicon.py
@@ -366,6 +366,19 @@ def test_get_lexicon_term_by_id(lexicon_term):
     )
 
 
+def test_get_lexicon_terms_sort_categories_branch():
+    """
+    Ensure the special-case branch (sort == 'categories') in GET /lexicon is exercised.
+    It should not apply sorting/filtering and still return a valid pagination payload.
+    """
+    resp = client.get("/lexicon/term", params={"sort": "categories"})
+    assert resp.status_code == 200
+    data = resp.json()
+    # fastapi-pagination returns a Page-like object with these keys
+    assert "items" in data
+    assert "total" in data
+
+
 def test_get_lexicon_term_by_id_404_not_found(lexicon_term):
     bad_id = 999999
     response = client.get(f"/lexicon/term/{bad_id}")
